Revert "fix: cast lastRun to int before passing to date()"
This reverts commit 3daf5e7fbce4804f2bab71e5c73ec727658050f0.

+++ src/ZeroSense/Features/Utilities/OrderDiagnostics.php
@@ -64,7 +64,7 @@ class OrderDiagnostics implements FeatureInterface
             return;
         }
 
-        $lastRun = (int) get_option('zs_order_diagnostics_last_run', 0);
+        $lastRun = get_option('zs_order_diagnostics_last_run', 0);
         $nonce   = wp_create_nonce('zs_order_diagnostics');
         ?>
         <div class="notice notice-info" style="display:flex; align-items:center; gap:12px;">
